moved routes from app.py to files grouped by their route uses. UserReportsService.py and UserSessionService.py. Renamed app.py to server.py. Able to call all services in postman without anything breaking. gg
