+++ backend-services/extensions.py
@@ -1,0 +1,5 @@
+from flask_pymongo import PyMongo
+from flask_mysqldb import MySQL
+
+mysql = MySQL()
+mongo = PyMongo()
